move charge to "common" and eliminate "local_path".
