+++ controller/api.py
@@ -1,29 +1,23 @@
 from flask import Flask, jsonify
 import requests
-import sys
 import threading
 import time
 import os
 import datetime
 
 from common import config
+from ctrldbops import get_hetzner, find_hosts, get_bitbsd, check_paid, deduct_host, get_suspended, delete_host
+from orchestrator import del_server
+
+# charge.py
+#invoice(msat=None, amount=0, cur='EUR', desc=False)
+#register_webhook(invoice_id, callback_url):
+from common.charge import invoice, register_webhook, get_invoice
 
 app = Flask(__name__)
 
 app.config['JSONIFY_PRETTYPRINT_REGULAR'] = True
 app.url_map.strict_slashes = False
-
-wallet_host = config['wallet']['host']
-project_path = config['paths']['local_path']
-sys.path.insert(1, project_path + '/wallet')
-
-# charge.py
-#invoice(msat=None, amount=0, cur='EUR', desc=False)
-#register_webhook(invoice_id, callback_url):
-from charge import invoice, register_webhook, get_invoice
-from ctrldbops import get_hetzner, find_hosts, get_bitbsd, check_paid, deduct_host, get_suspended, delete_host
-from orchestrator import del_server
-
 
 def accountant():
     threading.Timer(3600.0, accountant).start()
